Refactor addItem to use requestCart
When adding an item, we cannot safely pass a `ResponseCartProduct`
because we don't have the data needed yet to properly fill out the
properties; in fact, all we really have is a `RequestCartProduct`.

Furthermore, the way that the cart manager submits updates is not
sustainable for adding items. Prior to this change, an action causes the
stored `ResponseCart` to be modified and then sends it to the server,
replacing the `ResponseCart` with the result. However, this is
optimistic updating: the UI will react and render itself with the
requested data before the server request is made. This has been ok for
removing items, but when adding items we have an invalid `ResponseCart`
and the UI will throw errors with the invalid data.

This change modifies the way that an invalid cart is sent to the server.
We add a new property to the internal state called `requestCart` which
is used to stage changes before they are sent to the server without
modifying the current `responseCart` (until the change is applied by the
server and returned). When adding an item, we add the item to this
property (which conveniently is a `RequestCart` and so accepts the
`RequestCartProduct` that we have).

The other cart update actions should also be modified to use this
system, but here we just mark them with TODOs.